+++ src/main/java/com/belykh/finalProj/service/LotService.java
@@ -5,6 +5,7 @@ import com.belykh.finalProj.entity.LotHeader;
 import com.belykh.finalProj.entity.dbo.LotState;
 import com.belykh.finalProj.exception.ServiceException;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -18,8 +19,8 @@ public interface LotService {
     List<LotHeader> findLotHeadersByStateAndBuyerId(Long buyerId,LotState state) throws ServiceException;
     LotFull findFullLotInfo(Long id) throws ServiceException;
     boolean deleteLot(Long id,Long ownerId) throws ServiceException;
-    boolean buyLot(Long id,Long buyerId,Double price) throws ServiceException;
+    boolean buyLot(Long id, Long buyerId, BigDecimal price) throws ServiceException;
     boolean payLot(Long lotId) throws ServiceException;
-    boolean offerLot(Long ownerId, Long flowerId, Long cityId, String street, Integer houseNumber, double price, Integer count, LocalDateTime end, String description) throws ServiceException;
+    boolean offerLot(Long ownerId, Long flowerId, Long cityId, String street, Integer houseNumber, BigDecimal price, Integer count, LocalDateTime end, String description) throws ServiceException;
     boolean approveLot(Long lotId) throws ServiceException;
 }
